RF: Move internal clone helper to clone.py
